Move dateConverter to common utils namespace
[MAILPOET-2665]
